Split out HTTP utils in lib/url.
No code changes to any of the utils, only some moving around.
Splits out the tests too.
