+++ client/lib/url/index.ts
@@ -9,10 +9,14 @@ import { has, isString, omit, startsWith } from 'lodash';
  */
 import config from 'config';
 import { isLegacyRoute } from 'lib/route/legacy-routes';
-import { URL, SiteSlug, Scheme } from 'types';
+import { URL, Scheme } from 'types';
 import { Falsey } from 'utility-types';
 
+/**
+ * Re-exports
+ */
 export { addQueryArgs } from 'lib/route';
+export { withoutHttp, urlToSlug, urlToDomainAndPath } from './http-utils';
 
 /**
  * Check if a URL is located outside of Calypso.
@@ -66,27 +70,6 @@ export function isHttps( url: URL ): boolean {
 }
 
 const schemeRegex = /^\w+:\/\//;
-const urlWithoutHttpRegex = /^https?:\/\//;
-
-/**
- * Returns the supplied URL without the initial http(s).
- * @param  url The URL to remove http(s) from
- * @return     URL without the initial http(s)
- */
-export function withoutHttp( url: '' ): '';
-export function withoutHttp( url: Falsey ): null;
-export function withoutHttp( url: URL ): URL;
-export function withoutHttp( url: URL | Falsey ): URL | null {
-	if ( url === '' ) {
-		return '';
-	}
-
-	if ( ! url ) {
-		return null;
-	}
-
-	return url.replace( urlWithoutHttpRegex, '' );
-}
 
 export function addSchemeIfMissing( url: URL, scheme: Scheme ): URL {
 	if ( false === schemeRegex.test( url ) ) {
@@ -107,28 +90,6 @@ export function setUrlScheme( url: URL, scheme: Scheme ) {
 	}
 
 	return url.replace( schemeRegex, schemeWithSlashes );
-}
-
-export function urlToSlug( url: Falsey ): null;
-export function urlToSlug( url: URL ): SiteSlug;
-export function urlToSlug( url: URL | Falsey ): SiteSlug | null {
-	if ( ! url ) {
-		return null;
-	}
-
-	return withoutHttp( url ).replace( /\//g, '::' );
-}
-
-/**
- * Removes the `http(s)://` part and the trailing slash from an URL.
- * "http://blog.wordpress.com" will be converted into "blog.wordpress.com".
- * "https://www.wordpress.com/blog/" will be converted into "www.wordpress.com/blog".
- *
- * @param  urlToConvert The URL to convert
- * @return              The URL's domain and path
- */
-export function urlToDomainAndPath( urlToConvert: URL ): URL {
-	return withoutHttp( urlToConvert ).replace( /\/$/, '' );
 }
 
 /**
