+++ client/lib/url/package.json
@@ -1,0 +1,3 @@
+{
+	"sideEffects": false
+}
